bug/remove_type_checking
close #113
