--- changes/+loader-consolidation.refactor.md
@@ -1,1 +1,0 @@
-Internal: consolidated all loading machinery into the public ``Loader`` class. Function-form is ``Loader(*sources, schema=Cls, ...).load()``; decorator-form is the new ``Loader.as_decorator(*sources, ...)`` classmethod. ``dature.load(...)`` remains the recommended user-facing shortcut for both. ``loading/single.py`` removed (single-source report factory moved to ``load_report.py``, logger to ``loader.py``). ``loading/multi.py`` renamed to ``loading/merge.py`` and stripped to merge mechanics. ``LoadReport`` / ``FieldOrigin`` / ``SourceEntry`` extracted into ``dature/report_types.py`` to break the import cycle with ``masking`` (re-exported from ``load_report`` for compatibility). No public API change for end users.
