+++ docs/features/cli.md
@@ -218,4 +218,4 @@ The CLI exposes the common case. [`load()`](../api-reference.md#datureload) para
 
 ## Implementation note: dogfooding
 
-The `dature` console script parses its own arguments through [`ArgparseSource`](../api-reference.md#argparsesourceflatkeysource) — the very class users instantiate to feed CLI arguments into their own configurations. The dataclass schema that the CLI loads into is built at runtime from the signature of [`load()`](../api-reference.md#datureload) (via `dataclasses.make_dataclass` over `typing.get_type_hints(load)`), so adding or changing a parameter on `load()` is automatically reflected in the CLI without any hand-written wiring. The CLI is not special code: it is dature using itself.
+The `dature` console script parses its own arguments through [`ArgparseSource`](cli_source.md) — the very class users instantiate to feed CLI arguments into their own configurations. The dataclass schema that the CLI loads into is built at runtime from the signature of [`load()`](../api-reference.md#datureload) (via `dataclasses.make_dataclass` over `typing.get_type_hints(load)`), so adding or changing a parameter on `load()` is automatically reflected in the CLI without any hand-written wiring. The CLI is not special code: it is dature using itself.
