+++ examples/docs/advanced/merge_rules/advanced_merge_rules_custom_source.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from pathlib import Path
 
 import dature
-from dature.strategies.source import LoadCtx, SourceMergeStrategy
+from dature.loading.merge_runtime import LoadCtx, SourceMergeStrategy
 from dature.types import JSONValue
 
 SHARED_DIR = Path(__file__).parents[2] / "shared"
