+++ src/dature/errors/__init__.py
@@ -1,19 +1,17 @@
 from dature.errors.exceptions import (
-    CaretSpan,
     DatureConfigError,
     DatureError,
     EnvVarExpandError,
     FieldGroupError,
     FieldGroupViolationError,
     FieldLoadError,
-    LineRange,
     MergeConflictError,
     MergeConflictFieldError,
     MissingEnvVarError,
     SourceLoadError,
-    SourceLocation,
     ValidatorTypeError,
 )
+from dature.errors.loc_types import CaretSpan, LineRange, SourceLocation
 
 __all__ = [
     "CaretSpan",
