+++ src/dature/loaders/json5_.py
@@ -1,8 +1,8 @@
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
+try:
     from json5 import JsonIdentifier
+except ImportError:  # pragma: no cover  -- ``json5`` extra not installed
+    JsonIdentifier = str  # type: ignore[misc, assignment]
 
 
-def str_from_json_identifier(value: "JsonIdentifier") -> str:
+def str_from_json_identifier(value: JsonIdentifier) -> str:
     return str(value)
