+++ src/dature/sources/ini_.py
@@ -7,8 +7,7 @@ from adaptix.provider import Provider
 
 from dature.expansion.env_expand import expand_env_vars
 from dature.path_finders.ini_ import TablePathFinder
-from dature.sources.base import FileSource
-from dature.sources.retort import string_value_loaders
+from dature.sources.base import FileSource, string_value_loaders
 from dature.types import BINARY_IO_TYPES, TEXT_IO_TYPES, ExpandEnvVarsMode, FileOrStream, JSONValue
