+++ src/dature/validators/base.py
@@ -1,5 +1,5 @@
 from collections.abc import Iterable, Mapping
-from typing import TYPE_CHECKING, Annotated, Any, cast, get_args, get_origin, get_type_hints
+from typing import Annotated, Any, cast, get_args, get_origin, get_type_hints
 
 from adaptix import P, validator
 from adaptix.load_error import AggregateLoadError, ValidationLoadError
@@ -8,13 +8,11 @@ from adaptix.struct_trail import append_trail
 
 from dature.expansion.alias_provider import resolve_nested_owner
 from dature.field_path import FieldPath
-from dature.types import FieldValidators
+from dature.protocols import DataclassInstance
 from dature.validators.collection import EachPredicate
 from dature.validators.predicate import AndPredicate, Predicate
 from dature.validators.root import RootPredicate
-
-if TYPE_CHECKING:
-    from dature.protocols import DataclassInstance
+from dature.validators.types import FieldValidators
 
 
 def _flatten(predicate: Predicate) -> list[Predicate]:
