+++ src/dature/validators/types.py
@@ -1,0 +1,4 @@
+from dature.types import FieldRef
+from dature.validators.predicate import Predicate
+
+type FieldValidators = dict[FieldRef, Predicate | tuple[Predicate, ...]]
